a little more cleanup

+++ lib/includes/nghttp2/windows/nghttp2_windows.h
@@ -1,7 +1,6 @@
 #ifndef _NGHTTP2_WINDOWS_H_
 #define _NGHTTP2_WINDOWS_H_
 
-#include <basetsd.h>
 #include <WinSock2.h>
 
 #ifndef _SSIZE_T
@@ -10,7 +9,7 @@
 /* posix defines ssize_t as used for byte count and error indication.
  * it is not part of the C standard and missing in Visual Studio.
  */
-typedef SSIZE_T ssize_t;
+typedef long ssize_t;
 #endif /* _SSIZE_T */
 
 #ifndef _ATTRIBUTE_UNUSED
